Add support for database selection at runtime
This is implemented via DBFactory. Repurpose the --db-path runtime parameter for selecting the db path and add a new parameter (--data-dir) for setting the configuration file / keystore path, and remove the Defaults class (which was used to store the set DB path) since that functionality has been replaced with DBFactory.