Support RocksDB as a database backend
Add support for RocksDB to both Aleth and TestEth. RocksDB is included via the Hunter package manager and is compiled into Aleth and TestEth. One can select RocksDB at runtime via the '--db' command-line argument. Note that Aleth still defaults to LevelDB and TestEth still defaults to MemoryDB. The plan is to perform more testing on RocksDB and then switch it on as default (and remove LevelDB support).