Add FindNode test cases
Add 2 tests - the first validates that a Neighbours packet is sent in response to a FindNode packet from a node for which the endpoint proof has been completed. The second verifies that a Neighbours packet isn't sent in response to an unsolicited FindNode packet (i.e. a FindNode packet from a node for which an endpoint proof hasn't been completed).

Note that adding these tests also requires modifying the concurrent queue's pop() function to accept a wait timeout since the second test needs to wait for the Neighbours packet for a finite amount of time (since it is never expected to arrive)