Overload concurrent queue pop
Overload pop() so it can be called without a timeout - this enables us to keep its old behavior (wait indefinitely until the condition variable is woken up) while supporting the new behavior (wait with timeout) for tests which require it.

Also untabify the source file, auto-format it with clang-format, and update the license message.